+++ sync-for-payroll/src/codatsyncpayroll/sdk.py
@@ -12,6 +12,7 @@ from .sdkconfiguration import SDKConfiguration
 from .tracking_categories import TrackingCategories
 from codatsyncpayroll import utils
 from codatsyncpayroll.models import shared
+from typing import Dict
 
 class CodatSyncPayroll:
     r"""Sync for Payroll: The API for Sync for Payroll.
@@ -58,7 +59,7 @@ class CodatSyncPayroll:
                  security: shared.Security = None,
                  server_idx: int = None,
                  server_url: str = None,
-                 url_params: dict[str, str] = None,
+                 url_params: Dict[str, str] = None,
                  client: requests_http.Session = None,
                  retry_config: utils.RetryConfig = None
                  ) -> None:
@@ -71,7 +72,7 @@ class CodatSyncPayroll:
         :param server_url: The server URL to use for all operations
         :type server_url: str
         :param url_params: Parameters to optionally template the server URL with
-        :type url_params: dict[str, str]
+        :type url_params: Dict[str, str]
         :param client: The requests.Session HTTP client to use for all operations
         :type client: requests_http.Session
         :param retry_config: The utils.RetryConfig to use globally
@@ -80,7 +81,9 @@ class CodatSyncPayroll:
         if client is None:
             client = requests_http.Session()
         
+        
         security_client = utils.configure_security_client(client, security)
+        
         
         if server_url is not None:
             if url_params is not None: